# Conflicts:
#	Helpie2/app/src/main/java/com/example/helpie/UiState.kt

+++ Helpie2/app/src/main/java/com/example/helpie/tripPlanificator/data/dto/response/ContinuousLegDto.kt
@@ -28,6 +28,6 @@ data class ContinuousLegDto(
     val duration: String,
 
     @PropertyElement(name = "ojp:Length")
-    val length: Int,
+    val length: Double,
 
 )
